[lib] Check mode+ownership in capabilities inspection
The capabilities inspection currently does not check file permissions
and ownership, so a world-writable binary with capabilities could pass
the inspection. To catch such cases, the capabilities inspection should
also check file modes and ownership, even when the permissions and
ownership inspections themselves are not being run.

Resolves https://github.com/rpminspect/rpminspect/issues/1040
