+++ lib/meson.build
@@ -89,12 +89,14 @@ librpminspect_sources = [
     'output_summary.c',
     'output_text.c',
     'output_xunit.c',
+    'ownership.c',
     'parse_dson.c',
     'parse_json.c',
     'parse_yaml.c',
     'pairfuncs.c',
     'paths.c',
     'peers.c',
+    'permissions.c',
     'readelf.c',
     'readfile.c',
     'rebase.c',
